+++ R/MetaDataDiagnostics.R
@@ -137,6 +137,12 @@ saveDatabaseMetaData <- function(databaseId,
     vocabularyVersion = !!vocabularyVersion,
     isMetaAnalysis = 0
   )
+  database <- makeDataExportable(
+    x = database,
+    tableName = "database",
+    databaseId = databaseId,
+    minCellCount = minCellCount
+  )
   writeToCsv(data = database,
              fileName = file.path(exportFolder, "database.csv"))
   delta <- Sys.time() - startMetaData
@@ -146,44 +152,6 @@ saveDatabaseMetaData <- function(databaseId,
     attr(delta, "units")
   ))
 }
-
-getCdmVocabularyVersion <- function(connection, cdmDatabaseSchema) {
-  vocabularyVersionCdm <- NULL
-  tryCatch({
-    vocabularyVersionCdm <-
-      DatabaseConnector::renderTranslateQuerySql(
-        connection = connection,
-        sql = "select * from @cdm_database_schema.cdm_source;",
-        cdm_database_schema = cdmDatabaseSchema,
-        snakeCaseToCamelCase = TRUE
-      ) %>%
-      dplyr::tibble()
-  }, error = function(...) {
-    warning("Problem getting vocabulary version. cdm_source table not found in the database.")
-    if (connection@dbms == "postgresql") { #this is for test that automated testing purpose
-      DatabaseConnector::dbExecute(connection, "ABORT;")
-    }
-  })
-
-  if (all(!is.null(vocabularyVersionCdm),
-          nrow(vocabularyVersionCdm) > 0,
-          'vocabularyVersion' %in% colnames(vocabularyVersionCdm))) {
-    if (nrow(vocabularyVersionCdm) > 1) {
-      warning('Please check ETL convention for OMOP cdm_source table. It appears that there is more than one row while only one is expected.')
-    }
-    vocabularyVersionCdm <- vocabularyVersionCdm %>%
-      dplyr::rename(vocabularyVersionCdm = .data$vocabularyVersion) %>%
-      dplyr::pull(vocabularyVersionCdm) %>%
-      max() %>%
-      unique()
-  } else {
-    warning("Problem getting vocabulary version. cdm_source table either does not have data, or does not have the field vocabulary_version.")
-    vocabularyVersionCdm <- "Unknown"
-  }
-
-  return(vocabularyVersionCdm)
-}
-
 
 getVocabularyVersion <- function(connection, vocabularyDatabaseSchema) {
   DatabaseConnector::renderTranslateQuerySql(