+++ NEWS.md
@@ -1,3 +1,48 @@
+CohortDiagnostics 3.0.0
+=======================
+Changes:
+
+1. Time series diagnostics removed
+
+2. Removed runCohortDiagnostics function - this has now been completely replaced with executeDiagnostics
+
+3. Removed `loadCohortsFromPackage` function as this is now replaced with
+`CohortGenerator::getCohortDefinitionSet`
+
+4. Removed instantiate cohort functionality, `instantiateCohortSet` should now be used with the `CohortGenerator` package
+
+5. Removed optional `inclusionStatisticsFolder` parameter, this is now all exported directly from `CohortGenerator` 
+without the need to generate this first.
+
+6. Removed usage of Rdata files in DiagnosticsExplorer shiny app and function to create them `preMergeDiagnosticsFiles`
+
+7. Added function `createMergedResultsFile` which outputs a shiny app
+
+8. Added support for any `SqlRender/DatabaseConnector` compatible database (note, 
+this is experimental. Postgres and sqlite are the only backends recommended for use in production environments)
+
+9. Improved metadata collection and storage from runs of cohort diagnostics.
+
+10. Removed phenotype_id field from data ddl
+
+11. Additional checks to the output of cohort diagnostics to ensure it conforms to its own results data model. The new function (internal) is makeDataExportable. Results data model csv file has been enhanced with new fields, including a field to specify if the value is to be subjected to privacy protection (i.e. min cell count, eg. person count). Note a bug was discovered in the orphan concepts and included source concepts that was leading to duplication of row records by primary key. This bug has been fixed by calculating its max value grouped by primary keys. It will be fixed in another commit.
+
+CohortDiagnostics 2.2.3
+=======================
+
+Bug fixes:
+
+1. Replace use dplyr across() for bug introduced by tidyr v1.2.0
+
+
+CohortDiagnostics 2.2.2
+=======================
+
+Bug fixes:
+
+1. Fixed syntax error causing empty description field in inclusion rule stats
+
+
 CohortDiagnostics 2.2.1
 =======================